6 Juni 2018

+++ database/migrations/2018_06_05_092956_create_hasilbobots_table.php
@@ -15,7 +15,7 @@ class CreateHasilbobotsTable extends Migration
     {
         Schema::create('hasilbobots', function (Blueprint $table) {
             $table->bigInteger('nim',20)->unsigned();
-            $table->Integer('nilai');
+            $table->decimal('nilai',8, 4);
             $table->timestamps();
         });
     }
